backend/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
click==8.1.7
fastapi==0.115.4
h11==0.14.0
idna==3.10
Jinja2==3.1.4
MarkupSafe==3.0.2
msgpack==1.1.0
pydantic==2.9.2
pydantic_core==2.23.4
python-multipart==0.0.17
python-nmap==0.7.1
sniffio==1.3.1
starlette==0.41.2
typing_extensions==4.12.2
uvicorn==0.32.0
